move src/configs.ts to lib directory
