+++ src/run.ts
@@ -10,7 +10,7 @@ import {
   BurnDynamicProofConfig,
   TransferDynamicProofConfig,
   UpdatesDynamicProofConfig,
-} from './configs.js';
+} from './lib/configs.js';
 import {
   generateDummyDynamicProof,
   program,
